backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv==1.0.1
